#852: Fix the NoNulls classes API inconsistency
